remove rtmidi.dll from the NuGet package.
The situation is kinda cosmetic yet complicated; if we add dll under lib/net45
it will be resolved and added as a <Reference> to a referencing project.
It seems harmless, our C# compiler ignores it, but it is awkward to have an
unmanaged dll as a <Reference> and being added to the project.