Now MidiPlayer is free from internal real time tracking.
To make it really independent of "real time", any references to
DateTime.Now has to be eliminated. It has been tracked for tempo
adjustment for old code, but they are gone.

This however exposed some issue in unit tests which actually did not
work well without real time yet. It is due to player's threading
(it is concurrent and thus it involves indeterminant thread context
switching). After some investigation, I noticed that those events
needs to be ensured to have been fired (when testing).

It is ugly, but due to that reason, there is now Task.Delay() until
the expected event happens and player makes progress after some
thread context switching.