Fixed MidiEvent conversion that dropped sysex status byte https://github.com/atsushieno/managed-midi/issues/59
According to my github activities, it is most likely that commit f92f7c44
which introduced this was introduced when I was dealing with fluidsynth API
(I also have NFluidsynth.MidiAccess) which did not require status byte
when it takes sysex bytes.

It somehow does not break sysex messaging in ALSA so it was left unnoticed.